Refactored code to make simpler for users to import. May eventually need refactoring into folders again, but this is good for the timebeing.
